OO version of the git
